requirements.txt
# Core dependencies first
pydantic
pydantic-settings

# FastAPI & Web Framework
fastapi
uvicorn[standard]

# LangChain ecosystem
langchain-core
langchain-google-genai

# Google Gemini Integration
google-generativeai

langchain
langgraph

# Data Processing
python-dotenv

# Utilities
httpx